Moved EventConstants and TelemetryConstants to core
Needed in core to be able to use the endpoint names from other
services. Used in AbstractVertxBasedHttpProtocolAdapter.
